requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
magic-filter==1.0.12
multidict==6.1.0
openai==1.58.1
propcache==0.2.1
protobuf==5.29.2
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
whatsapp-api-client-python==0.0.44
whatsapp-chatbot-python==0.8.0
yandex-speechkit==1.5.0
yarl==1.18.3
